AppController: added confirmation form for trip and plan delete, added js trip confirmDeleteTrip.jsp: changed to spring form - auto crsf token confirmDeletePlan.jsp: changed to spring form - auto crsf token RestServiceController: added controller for js form fetch queries SecurityConfig: added authorization for /rest/** TripForm.js: added tripForm.jsp: added tripFormStyle.css: added

+++ src/main/java/pl/kamfonik/bytheway/controller/AppController.java
@@ -249,30 +249,28 @@ public class AppController {
                                     @PathVariable Long id){
         User user = currentUser.getUser();
         if (user.isAdmin() || tripService.checkUserTrip(id,user)) {
-//            model.addAttribute("trip",tripService.findTripById(id));
-//            return "/app/confirmDeleteTrip";
-            tripService.delete(id);
-            return "redirect:/app";
+            model.addAttribute("trip",tripService.findTripById(id));
+            return "/app/confirmDeleteTrip";
         } else {
             return "redirect:/403";
         }
     }
 
-//    @PostMapping("/delete/trip/{id}")
-//    public String deleteTrip(@AuthenticationPrincipal CurrentUser currentUser,
-//                             @PathVariable Long id,
-//                             @RequestParam String choice){
-//        if("No".equals(choice)){
-//            return "redirect:/app";
-//        }
-//        User user = currentUser.getUser();
-//        if (user.isAdmin() || tripService.checkUserTrip(id,user)) {
-//            tripService.delete(id);
-//        } else {
-//            return "redirect:/403";
-//        }
-//        return "redirect:/app";
-//    }
+    @PostMapping("/delete/trip/{id}")
+    public String deleteTrip(@AuthenticationPrincipal CurrentUser currentUser,
+                             @PathVariable Long id,
+                             @RequestParam String choice){
+        if("No".equals(choice)){
+            return "redirect:/app";
+        }
+        User user = currentUser.getUser();
+        if (user.isAdmin() || tripService.checkUserTrip(id,user)) {
+            tripService.delete(id);
+        } else {
+            return "redirect:/403";
+        }
+        return "redirect:/app";
+    }
 
     @GetMapping("/delete/plan/{id}")
     public String confirmDeletePlan(Model model,
@@ -280,28 +278,31 @@ public class AppController {
                                     @PathVariable Long id){
         User user = currentUser.getUser();
         if (user.isAdmin() || planService.checkUserPlan(id,user)) {
-//            model.addAttribute("plan",planService.findPlanById(id));
-//            return "/app/confirmDeletePlan";
-            planService.delete(id);
-            return "redirect:/app";
+            model.addAttribute("plan",planService.findPlanById(id));
+            return "/app/confirmDeletePlan";
         } else {
             return "redirect:/403";
         }
     }
 
-//    @PostMapping("/delete/plan/{id}")
-//    public String deletePlan(@AuthenticationPrincipal CurrentUser currentUser,
-//                             @PathVariable Long id,
-//                             @RequestParam String choice){
-//        if("No".equals(choice)){
-//            return "redirect:/app";
-//        }
-//        User user = currentUser.getUser();
-//        if (user.isAdmin() || planService.checkUserPlan(id,user)) {
-//            planService.delete(id);
-//        } else {
-//            return "redirect:/403";
-//        }
-//        return "redirect:/app";
-//    }
+    @PostMapping("/delete/plan/{id}")
+    public String deletePlan(@AuthenticationPrincipal CurrentUser currentUser,
+                             @PathVariable Long id,
+                             @RequestParam String choice){
+        if("No".equals(choice)){
+            return "redirect:/app";
+        }
+        User user = currentUser.getUser();
+        if (user.isAdmin() || planService.checkUserPlan(id,user)) {
+            planService.delete(id);
+        } else {
+            return "redirect:/403";
+        }
+        return "redirect:/app";
+    }
+
+    @GetMapping("/add-trip")
+    public String addTripForm() {
+        return "app/tripForm";
+    }
 }
